requirements.txt
# Core Web Framework
fastapi>=0.100.0
uvicorn>=0.22.0
gunicorn>=20.1.0

# Rate Limiting
slowapi>=0.1.5

# Authentication & Security
python-jose>=3.3.0
pyjwt>=2.6.0
python-multipart>=0.0.6
cryptography>=40.0.0
certifi>=2023.5.7
bcrypt>=4.0.1

# Configuration & Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
tenacity>=8.2.0

# Redis Caching
redis>=4.5.0

# AI/ML Services
google-generativeai>=0.3.0
openai>=1.0.0

# Google Cloud
google-cloud-secret-manager>=2.16.0
google-cloud-logging>=3.5.0
google-cloud-storage>=2.9.0

# Networking & API Clients
httpx>=0.24.0
requests>=2.28.0
websockets>=10.4

# Date & Time Handling
python-dateutil>=2.8.2
pytz>=2023.3

# Utilities & Performance
orjson>=3.8.0
backoff>=2.2.1
structlog>=23.1.0
